requirements.txt
matplotlib==3.8.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
python_igraph==0.11.4
scipy==1.13.0
tqdm==4.65.0